+++ packages/api/src/extrinsic/extensions/index.ts
@@ -1,4 +1,4 @@
 export * from './SignedExtension.js';
 export * from './known/index.js';
 export * from './ExtraSignedExtension.js';
-export * from './NoopSignedExtension.js';
+export * from './FallbackSignedExtension.js';
